feat(constants): introduce antlrQueryConstants for operators and key types
